+++ hail/python/hailtop/batch/__init__.py
@@ -3,8 +3,8 @@ import nest_asyncio
 from .batch import Batch
 from .batch_pool_executor import BatchPoolExecutor
 from .backend import LocalBackend, ServiceBackend, Backend
-from .utils import BatchException
-from .functions import _combine, concatenate, plink_merge
+from .exceptions import BatchException
+from .utils import concatenate, plink_merge
 
 __all__ = ['Batch',
            'LocalBackend',
@@ -12,7 +12,6 @@ __all__ = ['Batch',
            'Backend',
            'BatchException',
            'BatchPoolExecutor',
-           '_combine',
            'concatenate',
            'genetics',
            'plink_merge'
