+++ hail/python/hailtop/batch/batch.py
@@ -6,7 +6,7 @@ from typing import Optional, Dict, Union, List, Any, Set
 from hailtop.utils import secret_alnum_string
 
 from . import backend as _backend, job, resource as _resource  # pylint: disable=cyclic-import
-from .utils import BatchException
+from .exceptions import BatchException
 
 
 class Batch:
