+++ hail/python/hailtop/batch/exceptions.py
@@ -1,0 +1,5 @@
+
+class BatchException(Exception):
+    def __init__(self, msg=''):
+        self.msg = msg
+        super().__init__(msg)
